Changed MDImage class
Added typehints to the MDImage methods.

Created method to overwrite parent method create_automatic_link
because this is not supported for images. This is currently
the best alternative when working with inheritance.

The most secure way would be to separate the creation of automatic links
from the MDLink class. This way the create_automatic_link function would
not be inherited.